requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.4
    # via django
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
confusable-homoglyphs==3.2.0
    # via django-registration
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==3.4.7
    # via jwcrypto
django-filter==2.4.0
    # via -r requirements.in
django-fsm==2.7.1
    # via -r requirements.in
django-oauth-toolkit==1.5.0
    # via -r requirements.in
django-registration==3.1.2
    # via -r requirements.in
django==3.2
    # via
    #   -r requirements.in
    #   django-filter
    #   django-oauth-toolkit
    #   django-registration
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-yasg
drf-access-policy==0.9.0
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
idna==2.10
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
jwcrypto==0.8
    # via django-oauth-toolkit
markupsafe==1.1.1
    # via jinja2
oauthlib==3.1.0
    # via django-oauth-toolkit
packaging==20.9
    # via drf-yasg
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via
    #   drf-access-policy
    #   packaging
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via django
requests==2.25.1
    # via
    #   coreapi
    #   django-oauth-toolkit
ruamel.yaml.clib==0.2.2
    # via ruamel.yaml
ruamel.yaml==0.17.4
    # via drf-yasg
six==1.15.0
    # via django-oauth-toolkit
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.4
    # via requests
